[ONNX] Support per channel quantization
Extending the support for quantization with per channel quantization.
An extra attribute `axis` can be found for per channel quantized tensors,
most commonly in quantized weight of Convolution or Linear module.
The PR adds support to correctly parse the `axis` attribute, and map to
ONNX representation in `QuantizeLinear` and `DequantizeLinear`.

ghstack-source-id: ccb055be6ee045435af69b36fd384c7caf6e4650
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76002