[JIT] Reuse traced fn for jit opinfos
Previously, jit opinfos would only run the traced function once. This is a problem for NNC and NVFuser, where the fused implementation only runs on the second invocation.

This caches the traced function and calls the cached implementation, so that subsequent calls actually perform fusion and use the fused implementation.

ghstack-source-id: 968ded9a7cb7519b77137936ea80eee245c63dcb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/76000